add error toast type and conditional to Toast component to swap between success and error
