+++ src/app/components/Toast.tsx
@@ -1,0 +1,76 @@
+"use client";
+
+import React from "react";
+
+interface ToastProps {
+  message: string;
+  type: "success" | "error";
+  onClose: () => void;
+}
+
+const Toast: React.FC<ToastProps> = ({ message, type, onClose }) => {
+  const isError = type === "error";
+
+  const containerClass = `fixed top-6 left-1/2 transform -translate-x-1/2 flex items-center w-full max-w-md p-4 text-sm text-gray-500 bg-white rounded-lg shadow-sm z-50`;
+
+  const iconWrapperClass = `inline-flex items-center justify-center shrink-0 w-8 h-8 rounded-lg ${
+    isError ? "text-red-500 bg-red-100" : "text-green-500 bg-green-100"
+  }`;
+
+  const icon = isError ? (
+    <svg
+      className="w-5 h-5"
+      aria-hidden="true"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="currentColor"
+      viewBox="0 0 20 20"
+    >
+      <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5Zm3.707 11.793a1 1 0 1 1-1.414 1.414L10 11.414l-2.293 2.293a1 1 0 0 1-1.414-1.414L8.586 10 6.293 7.707a1 1 0 0 1 1.414-1.414L10 8.586l2.293-2.293a1 1 0 0 1 1.414 1.414L11.414 10l2.293 2.293Z" />
+    </svg>
+  ) : (
+    <svg
+      className="w-5 h-5"
+      aria-hidden="true"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="currentColor"
+      viewBox="0 0 20 20"
+    >
+      <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5Zm3.707 8.207-4 4a1 1 0 0 1-1.414 0l-2-2a1 1 0 0 1 1.414-1.414L9 10.586l3.293-3.293a1 1 0 0 1 1.414 1.414Z" />
+    </svg>
+  );
+
+  return (
+    <div className={containerClass} role="alert">
+      <div className={iconWrapperClass}>
+        {icon}
+        <span className="sr-only">{isError ? "Error" : "Success"} icon</span>
+      </div>
+      <div className="ms-3 text-sm font-normal">{message}</div>
+      <button
+        onClick={onClose}
+        type="button"
+        className="ms-auto -mx-1.5 -my-1.5 text-gray-400 hover:text-gray-900 rounded-lg p-1.5 hover:bg-gray-100 inline-flex items-center justify-center h-8 w-8"
+        aria-label="Close"
+      >
+        <span className="sr-only">Close</span>
+        <svg
+          className="w-3 h-3"
+          aria-hidden="true"
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 14 14"
+        >
+          <path
+            stroke="currentColor"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth="2"
+            d="M1 1l6 6m0 0l6 6M7 7l6-6M7 7L1 13"
+          />
+        </svg>
+      </button>
+    </div>
+  );
+};
+
+export default Toast;
